+++ indexer-api/Cargo.toml
@@ -20,6 +20,7 @@ async-graphql      = { workspace = true, features = [ "uuid" ] }
 async-graphql-axum = { workspace = true }
 async-stream       = { workspace = true }
 axum               = { workspace = true, features = [ "http2" ] }
+bech32             = { workspace = true }
 byte-unit-serde    = { workspace = true }
 chacha20poly1305   = { workspace = true, features = [ "std" ] }
 clap               = { workspace = true, features = [ "derive" ] }
