+++ indexer-common/Cargo.toml
@@ -16,7 +16,6 @@ rustdoc-args = [ "--cfg", "docsrs" ]
 
 [dependencies]
 async-nats                   = { workspace = true, optional = true }
-bech32                       = { workspace = true }
 chacha20poly1305             = { workspace = true, features = [ "std" ] }
 const-hex                    = { workspace = true, features = [ "serde" ] }
 derive_more                  = { workspace = true, features = [ "as_ref", "deref", "display", "from", "into" ] }
